requirements.txt
python-telegram-bot[job-queue]==21.5
requests==2.31.0
